update README, move publicKey and token clients to separate files, remove url mutation functions, remove createClient functions from connectClient to avoid circular dependencies
