+++ chain-connect/README.md
@@ -19,3 +19,15 @@ Run `nx build connect` to build the library.
 ## Running unit tests
 
 Run `nx test connect` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Example client usage
+
+A new GalachainConnectClient can be created and connected by instantiating a new client object:
+
+> const client = new GalachainConnectClient();
+> await client.connectToMetaMask();
+
+PublicKeyClient and TokenClient can be instantiated with the GalachainConnectClient and the url as arguments:
+
+> const publicKeyClient = new PublicKeyClient(this, url)
+> const tokenClient = new TokenClient(this, url);
